Reduce warnings in bdl on gcc10 {DRQS 171723333} (#4256)
* Empty change

* Remove clang warnings from bdlma_sequentialpool.t.cpp

* Remove empty change

* Remove conversion warnings from bdlb_algorithmworkaroundutil.t.cpp

* Remove warnings from bdlb_arrayutil

* Fix warnings in bdlb_bigendian.t.cpp

+++ groups/bdl/bdlb/bdlb_algorithmworkaroundutil.t.cpp
@@ -237,7 +237,7 @@ int main(int argc, char *argv[])
 
                 {
                     const int *result = Util::lowerBound(INPUT, END, VALUE);
-                    const int offset = result - INPUT;
+                    const long offset = result - INPUT;
 
                     ASSERTV(EXP_LOWER, offset, EXP_LOWER == offset);
                     if (result < END) {
@@ -255,7 +255,7 @@ int main(int argc, char *argv[])
                     const int *result =
                         Util::lowerBound(INPUT, END, VALUE, bsl::less<int>());
 
-                    const int offset = result - INPUT;
+                    const long offset = result - INPUT;
 
                     ASSERTV(EXP_LOWER, offset, EXP_LOWER == offset);
                     if (result < END) {
@@ -271,7 +271,7 @@ int main(int argc, char *argv[])
                 }
                 {
                     const int *result = Util::upperBound(INPUT, END, VALUE);
-                    const int offset = result - INPUT;
+                    const long offset = result - INPUT;
 
                     ASSERTV(EXP_UPPER, offset, EXP_UPPER == offset);
                     if (result < END) {
@@ -286,7 +286,7 @@ int main(int argc, char *argv[])
                     const int *result =
                         Util::upperBound(INPUT, END, VALUE, bsl::less<int>());
 
-                    const int offset = result - INPUT;
+                    const long offset = result - INPUT;
 
                     ASSERTV(EXP_UPPER, offset, EXP_UPPER == offset);
                     if (result < END) {
